+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed built-in transition effects not loading when multiple transitions are created
+
 ## [0.0.3] - 2023-08-30
 ### Added
 - Added Parameter type `image`